+++ docs/source/conf.py
@@ -26,7 +26,7 @@ copyright = '2023, Jean Landercy'
 author = 'Jean Landercy'
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.2"
+release = "0.1.4"
 
 
 # -- General to_dict ---------------------------------------------------
@@ -81,3 +81,8 @@ nbsphinx_execute_arguments = [
 html_scaled_image_link = False
 nbsphinx_input_prompt = '[%s]'
 nbsphinx_output_prompt = '[%s]'
+
+# https://nbsphinx.readthedocs.io/en/0.9.2/configuration.html
+suppress_warnings = [
+    'nbsphinx',
+]