+++ scifit/solvers/specials.py
@@ -5,7 +5,6 @@ from scifit.interfaces.generic import FitSolverInterface
 
 
 class DebyeInternalEnergyFitSolver(FitSolverInterface):
-
     @staticmethod
     def debye_integral(n):
         """
@@ -41,11 +40,12 @@ class DebyeInternalEnergyFitSolver(FitSolverInterface):
         :param T_D: Debye Temperature
         :return: Molecular Internal Energy
         """
-        return 3 * x[:, 0] * DebyeInternalEnergyFitSolver.debye_integral(3)(T_D / x[:, 0])
+        return (
+            3 * x[:, 0] * DebyeInternalEnergyFitSolver.debye_integral(3)(T_D / x[:, 0])
+        )
 
 
 class CrankEquationSolver:
-
     def __init__(self, alpha):
         self.alpha = alpha
         self.roots = {}
@@ -78,7 +78,6 @@ class CrankEquationSolver:
 
 
 class CrankDiffusion:
-
     def __init__(self, alpha=5.0, radius=1e-3, n=40):
         self.n = n
         self.alpha = alpha
@@ -106,7 +105,6 @@ class CrankDiffusion:
 
 
 class CrankDiffusionFitSolver(FitSolverInterface):
-
     _helper = CrankDiffusion()
 
     def __init__(self, alpha=3.9, radius=1.9e-3, **kwargs):
@@ -120,7 +118,7 @@ class CrankDiffusionFitSolver(FitSolverInterface):
         Solve the modified Crank diffusion problem for a solid sphere
 
         .. math::
-        
+
             \\gamma(t) = \\frac{C(t,R)}{C(0,R)} = \\frac{\\alpha}{1 + \\alpha} + 6 \\alpha \\sum\\limits_{i=1}^\\infty \\frac{\\exp \\left(-\\mathcal{D}\\frac{q_n^2 t}{R^2}\\right)}{9(\\alpha + 1) + \\alpha^2q_n^2}
 
         Where:
@@ -142,51 +140,37 @@ class CrankDiffusionFitSolver(FitSolverInterface):
         """
         return CrankDiffusionFitSolver._helper.objective(x[:, 0], Kp, D)
 
-#
-# class RaneyKetonDehydrogenationFitSolver(FitSolverInterface):
-#
-#     R = 8.31446261815324     # J/mol.K
-#     T0 = 292.05              # K
-#     p0 = 101600              # Pa
-#     V = 190e-6               # m3 of isopropanol
-#     m = 2.7677               # g of Raney Nickel
-#     rho = 785                # kg/m³
-#     M = 60.1                 # g/mol
-#     n0 = 1000 * rho * V / M  # mol
-#
-#     @staticmethod
-#     def _objective(n0, V):
-#
-#         def wrapped(xi, k1, k2):
-#             return (1 - k2) / k1 * (xi / V) - (k2 / k1) * (n0 / V) * np.log((n0 - xi) / n0)
-#
-#         return wrapped
-#
-#     objective = _objective(n0, V)
-#
-#     @staticmethod
-#     def model(x, k1, k2):
-#         """
-#
-#         .. math::
-#
-#             r = k_1 \\theta_A = k_1\\frac{aA}{1 + aA + bB + cC}
-#
-#         .. math::
-#
-#             r = \\frac{1}{V}\\frac{\\mathrm{d}\\xi}{\\mathrm{d}t} \\simeq k_1\\frac{aA}{aA + bB} = k\\frac{a(n_0 - \\xi)}{a(n_0 - \\xi) + b\\xi}
-#
-#         .. math::
-#
-#             \\int\\limits_0^\\xi\\left(1 + \\frac{b\\xi}{a(n_0-\\xi)}\\right)\\mathrm{d}\\xi = \\int\\limits_0^t k_1V\\mathrm{d}t
-#
-#         .. math::
-#
-#             \\left(1 - \\frac{b}{a}\\right)\\xi - n_0\\frac{b}{a}\\ln\\left|\\frac{n_0 - \\xi}{n_0}\\right| = k_1Vt
-#
-#         :param x:
-#         :param k1:
-#         :param k2:
-#         :return:
-#         """
-#         return RaneyKetonDehydrogenationFitSolver.objective(x[:, 0], k1, k2)
+
+class RaneyKetonDehydrogenationFitSolver(FitSolverInterface):
+    def __init__(self, n0=3.5, V=200e-6, **kwargs):
+        self.n0 = n0  # m3 of isopropanol
+        self.V = V  # mol
+        super().__init__(**kwargs)
+
+    def model(self, x, k1, k2):
+        """
+
+        .. math::
+
+            r = k_1 \\theta_A = k_1\\frac{aA}{1 + aA + bB + cC}
+
+        .. math::
+
+            r = \\frac{1}{V}\\frac{\\mathrm{d}\\xi}{\\mathrm{d}t} \\simeq k_1\\frac{aA}{aA + bB} = k\\frac{a(n_0 - \\xi)}{a(n_0 - \\xi) + b\\xi}
+
+        .. math::
+
+            \\int\\limits_0^\\xi\\left(1 + \\frac{b\\xi}{a(n_0-\\xi)}\\right)\\mathrm{d}\\xi = \\int\\limits_0^t k_1V\\mathrm{d}t
+
+        .. math::
+
+            \\left(1 - \\frac{b}{a}\\right)\\xi - n_0\\frac{b}{a}\\ln\\left|\\frac{n_0 - \\xi}{n_0}\\right| = k_1Vt
+
+        :param x:
+        :param k1:
+        :param k2:
+        :return:
+        """
+        return (1 - k2) / k1 * (x[:, 0] / self.V) - (k2 / k1) * (
+            self.n0 / self.V
+        ) * np.log((self.n0 - x[:, 0]) / self.n0)